+++ src/pages/SketchDetailPage/ui/MusicShape/MusicSphere/index.ts
@@ -1,0 +1,1 @@
+export { MusicSphere } from './MusicSphere';